+++ Examples/Resources/IntuneDeviceConfigurationDefenderForEndpointOnboardingPolicyWindows10/2-Update.ps1
@@ -22,7 +22,7 @@ Configuration Example
 
     node localhost
     {
-        IntuneDeviceConfigurationDefenderForEndpointOnboardingPolicyWindows10 'Example'
+        IntuneDeviceConfigurationDefenderOnboardingPolicyWindows10 'Example'
         {
             AdvancedThreatProtectionAutoPopulateOnboardingBlob = $True; # Updated Property
             AdvancedThreatProtectionOnboardingFilename         = "WindowsDefenderATP.onboarding";