+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceConfigurationDefenderOnboardingPolicyWindows10/MSFT_IntuneDeviceConfigurationDefenderOnboardingPolicyWindows10.psm1
@@ -1,4 +1,4 @@
-Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneDeviceConfigurationHealthMonitoringConfigurationPolicyWindows10'
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneDeviceConfigurationDefenderOnboardingPolicyWindows10'
 
 function Get-TargetResource
 {
@@ -8,18 +8,32 @@ function Get-TargetResource
     (
         #region resource generator code
         [Parameter()]
-        [ValidateSet('notConfigured', 'enabled', 'disabled')]
-        [System.String]
-        $AllowDeviceHealthMonitoring,
+        [System.Boolean]
+        $AdvancedThreatProtectionAutoPopulateOnboardingBlob,
 
         [Parameter()]
         [System.String]
-        $ConfigDeviceHealthMonitoringCustomScope,
+        $AdvancedThreatProtectionOffboardingBlob,
 
         [Parameter()]
-        [ValidateSet('undefined', 'healthMonitoring', 'bootPerformance', 'windowsUpdates', 'privilegeManagement')]
-        [System.String[]]
-        $ConfigDeviceHealthMonitoringScope,
+        [System.String]
+        $AdvancedThreatProtectionOffboardingFilename,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingBlob,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingFilename,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowSampleSharing,
+
+        [Parameter()]
+        [System.Boolean]
+        $EnableExpeditedTelemetryReporting,
 
         [Parameter()]
         [System.String]
@@ -76,7 +90,7 @@ function Get-TargetResource
         $AccessTokens
     )
 
-    Write-Verbose -Message "Getting configuration of the Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName}"
+    Write-Verbose -Message "Getting configuration of the Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName}"
 
     try
     {
@@ -102,81 +116,77 @@ function Get-TargetResource
 
             $getValue = $null
             #region resource generator code
-            if (-not [string]::IsNullOrEmpty($Id))
+            if (-not [System.String]::IsNullOrEmpty($Id))
             {
                 $getValue = Get-MgBetaDeviceManagementDeviceConfiguration -All -Filter "Id eq '$Id'" -ErrorAction SilentlyContinue
             }
 
             if ($null -eq $getValue)
             {
-                Write-Verbose -Message "Could not find an Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with Id {$Id}"
+                Write-Verbose -Message "Could not find an Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with Id {$Id}"
 
                 if (-Not [string]::IsNullOrEmpty($DisplayName))
                 {
                     $getValue = Get-MgBetaDeviceManagementDeviceConfiguration `
                         -All `
                         -Filter "DisplayName eq '$($DisplayName -replace "'", "''")'" `
-                        -ErrorAction SilentlyContinue
+                        -ErrorAction SilentlyContinue | Where-Object `
+                        -FilterScript { `
+                            $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.windowsDefenderAdvancedThreatProtectionConfiguration' `
+                    }
+                    if ($null -eq $getValue)
+                    {
+                        Write-Verbose -Message "Could not find an Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with DisplayName {$DisplayName}"
+                        return $nullResult
+                    }
+                    if (([array]$getValue).Count -gt 1)
+                    {
+                        throw "A policy with a duplicated displayName {'$DisplayName'} was found - Ensure displayName is unique"
+                    }
                 }
             }
             #endregion
-            if ($null -eq $getValue)
-            {
-                Write-Verbose -Message "Could not find an Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with DisplayName {$DisplayName}"
-                return $nullResult
-            }
         }
         else
         {
             $getValue = $Script:exportedInstance
         }
+
         $Id = $getValue.Id
-        Write-Verbose -Message "An Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName} was found."
-
-        #region resource generator code
-        $enumAllowDeviceHealthMonitoring = $null
-        if ($null -ne $getValue.AdditionalProperties.allowDeviceHealthMonitoring)
-        {
-            $enumAllowDeviceHealthMonitoring = $getValue.AdditionalProperties.allowDeviceHealthMonitoring.ToString()
-        }
-
-        $enumConfigDeviceHealthMonitoringScope = @()
-        if ($null -ne $getValue.AdditionalProperties.configDeviceHealthMonitoringScope)
-        {
-            $enumConfigDeviceHealthMonitoringScope = $getValue.AdditionalProperties.configDeviceHealthMonitoringScope.ToString().Split(',')
-
-        }
-        #endregion
+        Write-Verbose -Message "An Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName} was found."
 
         $results = @{
             #region resource generator code
-            AllowDeviceHealthMonitoring             = $enumAllowDeviceHealthMonitoring
-            ConfigDeviceHealthMonitoringCustomScope = $getValue.AdditionalProperties.configDeviceHealthMonitoringCustomScope
-            ConfigDeviceHealthMonitoringScope       = $enumConfigDeviceHealthMonitoringScope
-            Description                             = $getValue.Description
-            DisplayName                             = $getValue.DisplayName
-            Id                                      = $getValue.Id
-            RoleScopeTagIds                         = $getValue.RoleScopeTagIds
-            Ensure                                  = 'Present'
-            Credential                              = $Credential
-            ApplicationId                           = $ApplicationId
-            TenantId                                = $TenantId
-            ApplicationSecret                       = $ApplicationSecret
-            CertificateThumbprint                   = $CertificateThumbprint
-            ManagedIdentity                         = $ManagedIdentity.IsPresent
-            AccessTokens                            = $AccessTokens
+            AdvancedThreatProtectionAutoPopulateOnboardingBlob = $getValue.AdditionalProperties.advancedThreatProtectionAutoPopulateOnboardingBlob
+            AdvancedThreatProtectionOffboardingBlob            = $getValue.AdditionalProperties.advancedThreatProtectionOffboardingBlob
+            AdvancedThreatProtectionOffboardingFilename        = $getValue.AdditionalProperties.advancedThreatProtectionOffboardingFilename
+            AdvancedThreatProtectionOnboardingBlob             = $getValue.AdditionalProperties.advancedThreatProtectionOnboardingBlob
+            AdvancedThreatProtectionOnboardingFilename         = $getValue.AdditionalProperties.advancedThreatProtectionOnboardingFilename
+            AllowSampleSharing                                 = $getValue.AdditionalProperties.allowSampleSharing
+            EnableExpeditedTelemetryReporting                  = $getValue.AdditionalProperties.enableExpeditedTelemetryReporting
+            Description                                        = $getValue.Description
+            DisplayName                                        = $getValue.DisplayName
+            Id                                                 = $getValue.Id
+            RoleScopeTagIds                                    = $getValue.RoleScopeTagIds
+            Ensure                                             = 'Present'
+            Credential                                         = $Credential
+            ApplicationId                                      = $ApplicationId
+            TenantId                                           = $TenantId
+            ApplicationSecret                                  = $ApplicationSecret
+            CertificateThumbprint                              = $CertificateThumbprint
+            ManagedIdentity                                    = $ManagedIdentity.IsPresent
+            AccessTokens                                       = $AccessTokens
             #endregion
         }
-
-        $assignmentsValues = Get-MgBetaDeviceManagementDeviceConfigurationAssignment -DeviceConfigurationId $Id
-        $assignmentResult = @()
-        if ($assignmentsValues.Count -gt 0)
+        $returnAssignments = @()
+        $graphAssignments = Get-MgBetaDeviceManagementDeviceConfigurationAssignment -DeviceConfigurationId $Id
+        if ($graphAssignments.Count -gt 0)
         {
-            $assignmentResult += ConvertFrom-IntunePolicyAssignment `
+            $returnAssignments += ConvertFrom-IntunePolicyAssignment `
                 -IncludeDeviceFilter:$true `
-                -Assignments ($assignmentsValues)
+                -Assignments ($graphAssignments)
         }
-        $results.Add('Assignments', $assignmentResult)
+        $results.Add('Assignments', $returnAssignments)
 
         return $results
     }
@@ -199,18 +209,32 @@ function Set-TargetResource
     (
         #region resource generator code
         [Parameter()]
-        [ValidateSet('notConfigured', 'enabled', 'disabled')]
-        [System.String]
-        $AllowDeviceHealthMonitoring,
+        [System.Boolean]
+        $AdvancedThreatProtectionAutoPopulateOnboardingBlob,
 
         [Parameter()]
         [System.String]
-        $ConfigDeviceHealthMonitoringCustomScope,
+        $AdvancedThreatProtectionOffboardingBlob,
 
         [Parameter()]
-        [ValidateSet('undefined', 'healthMonitoring', 'bootPerformance', 'windowsUpdates', 'privilegeManagement')]
-        [System.String[]]
-        $ConfigDeviceHealthMonitoringScope,
+        [System.String]
+        $AdvancedThreatProtectionOffboardingFilename,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingBlob,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingFilename,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowSampleSharing,
+
+        [Parameter()]
+        [System.Boolean]
+        $EnableExpeditedTelemetryReporting,
 
         [Parameter()]
         [System.String]
@@ -232,7 +256,6 @@ function Set-TargetResource
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $Assignments,
         #endregion
-
         [Parameter()]
         [System.String]
         [ValidateSet('Absent', 'Present')]
@@ -280,15 +303,23 @@ function Set-TargetResource
     #endregion
 
     $currentInstance = Get-TargetResource @PSBoundParameters
-    $boundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+
+    $BoundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
 
     if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
-        Write-Verbose -Message "Creating an Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with DisplayName {$DisplayName}"
-        $boundParameters.Remove('Assignments') | Out-Null
+        Write-Verbose -Message "Creating an Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with DisplayName {$DisplayName}"
+        $BoundParameters.Remove('Assignments') | Out-Null
 
-        $CreateParameters = ([Hashtable]$boundParameters).Clone()
+        $CreateParameters = ([Hashtable]$BoundParameters).Clone()
         $CreateParameters = Rename-M365DSCCimInstanceParameter -Properties $CreateParameters
+
+        if ($AdvancedThreatProtectionAutoPopulateOnboardingBlob -and `
+                $PSBoundParameters.AdvancedThreatProtectionAutoPopulateOnboardingBlob)
+        {
+            $CreateParameters.Remove('AdvancedThreatProtectionOnboardingBlob') | Out-Null
+        }
+
         $CreateParameters.Remove('Id') | Out-Null
 
         $keys = (([Hashtable]$CreateParameters).Clone()).Keys
@@ -299,20 +330,8 @@ function Set-TargetResource
                 $CreateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters.$key
             }
         }
-		if ($CreateParameters.ContainsKey('ConfigDeviceHealthMonitoringScope')) {
-			$CreateParameters['configDeviceHealthMonitoringScope'] = ($CreateParameters['ConfigDeviceHealthMonitoringScope'] -join ',')
-			$CreateParameters.Remove('ConfigDeviceHealthMonitoringScope') | Out-Null
-		}
-		if ($CreateParameters.ContainsKey('AllowDeviceHealthMonitoring')) {
-			$CreateParameters['allowDeviceHealthMonitoring'] = $CreateParameters['AllowDeviceHealthMonitoring']
-			$CreateParameters.Remove('AllowDeviceHealthMonitoring') | Out-Null
-		}
-		if ($CreateParameters.ContainsKey('ConfigDeviceHealthMonitoringCustomScope')) {
-			$CreateParameters['configDeviceHealthMonitoringCustomScope'] = $CreateParameters['ConfigDeviceHealthMonitoringCustomScope']
-			$CreateParameters.Remove('ConfigDeviceHealthMonitoringCustomScope') | Out-Null
-		}
         #region resource generator code
-        $CreateParameters.Add('@odata.type', '#microsoft.graph.windowsHealthMonitoringConfiguration')
+        $CreateParameters.Add('@odata.type', '#microsoft.graph.windowsDefenderAdvancedThreatProtectionConfiguration')
         $policy = New-MgBetaDeviceManagementDeviceConfiguration -BodyParameter $CreateParameters
         $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $Assignments
 
@@ -326,11 +345,17 @@ function Set-TargetResource
     }
     elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Updating the Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with Id {$($currentInstance.Id)}"
-        $boundParameters.Remove('Assignments') | Out-Null
+        Write-Verbose -Message "Updating the Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with Id {$($currentInstance.Id)}"
+        $BoundParameters.Remove('Assignments') | Out-Null
 
-        $UpdateParameters = ([Hashtable]$boundParameters).Clone()
+        $UpdateParameters = ([Hashtable]$BoundParameters).Clone()
         $UpdateParameters = Rename-M365DSCCimInstanceParameter -Properties $UpdateParameters
+
+        if ($AdvancedThreatProtectionAutoPopulateOnboardingBlob -and `
+                $PSBoundParameters.AdvancedThreatProtectionAutoPopulateOnboardingBlob)
+        {
+            $UpdateParameters.Remove('AdvancedThreatProtectionOnboardingBlob') | Out-Null
+        }
 
         $UpdateParameters.Remove('Id') | Out-Null
 
@@ -342,20 +367,8 @@ function Set-TargetResource
                 $UpdateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters.$key
             }
         }
-		if ($UpdateParameters.ContainsKey('ConfigDeviceHealthMonitoringScope')) {
-			$UpdateParameters['configDeviceHealthMonitoringScope'] = ($UpdateParameters['ConfigDeviceHealthMonitoringScope'] -join ',')
-			$UpdateParameters.Remove('ConfigDeviceHealthMonitoringScope') | Out-Null
-		}
-		if ($UpdateParameters.ContainsKey('AllowDeviceHealthMonitoring')) {
-			$UpdateParameters['allowDeviceHealthMonitoring'] = $UpdateParameters['AllowDeviceHealthMonitoring']
-			$UpdateParameters.Remove('AllowDeviceHealthMonitoring') | Out-Null
-		}
-		if ($UpdateParameters.ContainsKey('ConfigDeviceHealthMonitoringCustomScope')) {
-			$UpdateParameters['configDeviceHealthMonitoringCustomScope'] = $UpdateParameters['ConfigDeviceHealthMonitoringCustomScope']
-			$UpdateParameters.Remove('ConfigDeviceHealthMonitoringCustomScope') | Out-Null
-		}
         #region resource generator code
-        $UpdateParameters.Add('@odata.type', '#microsoft.graph.windowsHealthMonitoringConfiguration')
+        $UpdateParameters.Add('@odata.type', '#microsoft.graph.windowsDefenderAdvancedThreatProtectionConfiguration')
         Update-MgBetaDeviceManagementDeviceConfiguration  `
             -DeviceConfigurationId $currentInstance.Id `
             -BodyParameter $UpdateParameters
@@ -368,7 +381,7 @@ function Set-TargetResource
     }
     elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Removing the Intune Device Configuration Health Monitoring Configuration Policy for Windows10 with Id {$($currentInstance.Id)}"
+        Write-Verbose -Message "Removing the Intune Device Configuration Defender For Endpoint Onboarding Policy for Windows10 with Id {$($currentInstance.Id)}"
         #region resource generator code
         Remove-MgBetaDeviceManagementDeviceConfiguration -DeviceConfigurationId $currentInstance.Id
         #endregion
@@ -383,18 +396,32 @@ function Test-TargetResource
     (
         #region resource generator code
         [Parameter()]
-        [ValidateSet('notConfigured', 'enabled', 'disabled')]
-        [System.String]
-        $AllowDeviceHealthMonitoring,
+        [System.Boolean]
+        $AdvancedThreatProtectionAutoPopulateOnboardingBlob,
 
         [Parameter()]
         [System.String]
-        $ConfigDeviceHealthMonitoringCustomScope,
+        $AdvancedThreatProtectionOffboardingBlob,
 
         [Parameter()]
-        [ValidateSet('undefined', 'healthMonitoring', 'bootPerformance', 'windowsUpdates', 'privilegeManagement')]
-        [System.String[]]
-        $ConfigDeviceHealthMonitoringScope,
+        [System.String]
+        $AdvancedThreatProtectionOffboardingFilename,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingBlob,
+
+        [Parameter()]
+        [System.String]
+        $AdvancedThreatProtectionOnboardingFilename,
+
+        [Parameter()]
+        [System.Boolean]
+        $AllowSampleSharing,
+
+        [Parameter()]
+        [System.Boolean]
+        $EnableExpeditedTelemetryReporting,
 
         [Parameter()]
         [System.String]
@@ -460,8 +487,10 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    $compareParameters = Get-CompareParameters
     $result = Test-M365DSCTargetResource -DesiredValues $PSBoundParameters `
-                                         -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '')
+                                             -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '') `
+                                             @compareParameters
     return $result
 }
 
@@ -525,7 +554,7 @@ function Export-TargetResource
         [array]$getValue = Get-MgBetaDeviceManagementDeviceConfiguration -Filter $Filter -All `
             -ErrorAction Stop | Where-Object `
             -FilterScript { `
-                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.windowsHealthMonitoringConfiguration' `
+                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.windowsDefenderAdvancedThreatProtectionConfiguration' `
         }
         #endregion
 
@@ -566,7 +595,8 @@ function Export-TargetResource
             }
 
             $Script:exportedInstance = $config
-            $Results = Get-TargetResource @Params
+            $Results = Get-TargetResource @params
+
             if ($Results.Assignments)
             {
                 $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName DeviceManagementConfigurationPolicyAssignments
@@ -613,5 +643,15 @@ function Export-TargetResource
         }
     }
 }
+function Get-CompareParameters
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable])]
+    param()
 
-Export-ModuleMember -Function *-TargetResource
+    return @{
+        ExcludedProperties = @('AdvancedThreatProtectionOnboardingBlob')
+    }
+}
+
+Export-ModuleMember -Function @('*-TargetResource', 'Get-CompareParameters')